Thank you for this opportunity!
Quick note if I had longer I would have added logging and made a more
robust way to iterate through the xml tree.

I would also move the more robust iterator to its own file and function
so that it could be reused for other projects.

There also appears to be an issue with people not filling out the
bedroom and bathrooms field.
Parsing the description appears to have correct numbers.

If I had more time I would use the data in the description to create a
better data set by parsing it.